feat: Add message sending status to Message type
Introduces a new `status` field to the `Message` type in `types.ts`. This field allows for tracking the sending state of messages ('sending', 'sent', 'error'), which can be used to provide UI feedback to the user in the chat interface.

The changes in `pages/Chat.tsx` are preparatory and pave the way for implementing this UI feedback mechanism.